feat: add connection metric
